Split out GetUserFromSession
