+++ vault/rotator.go
@@ -3,8 +3,6 @@ package vault
 import (
 	"fmt"
 	"log"
-	"regexp"
-	"strings"
 	"time"
 
 	"github.com/99designs/aws-vault/prompt"
@@ -162,39 +160,6 @@ func (r *Rotator) Rotate(profile string) error {
 
 	log.Printf("Rotated credentials for profile %q in vault", profile)
 	return nil
-}
-
-var (
-	getUserErrorRegexp = regexp.MustCompile(`^AccessDenied: User: arn:aws:iam::(\d+):user/(.+) is not`)
-)
-
-// GetUsernameFromSession returns the IAM username (or root) associated with the current aws session
-func GetUsernameFromSession(sess *session.Session) (string, error) {
-	client := iam.New(sess)
-
-	resp, err := client.GetUser(&iam.GetUserInput{})
-	if err != nil {
-		// Even if GetUser fails, the current user is included in the error. This happens when you have o IAM permissions
-		// on the master credentials, but have permission to use assumeRole later
-		matches := getUserErrorRegexp.FindStringSubmatch(err.Error())
-		if len(matches) > 0 {
-			pathParts := strings.Split(matches[2], "/")
-			return pathParts[len(pathParts)-1], nil
-		}
-
-		return "", err
-	}
-
-	if resp.User.UserName != nil {
-		return *resp.User.UserName, nil
-	}
-
-	if resp.User.Arn != nil {
-		arnParts := strings.Split(*resp.User.Arn, ":")
-		return arnParts[len(arnParts)-1], nil
-	}
-
-	return "", fmt.Errorf("Couldn't determine current username")
 }
 
 func retry(duration time.Duration, sleep time.Duration, callback func() error) (err error) {
